Record field counts, and null table manamagent tweaks.
it's possible for null tables, records, and table headers to dissagree on column count.
This can be down to columns being added to a table with existing data but where the records themselves are not updated.
It can be down to a column being deleted from the table, in these cases there is a still a reference on the records that were present before the removal.
etc.

Needed to confirm confirm field count from record unpack instead of relying on header.

Needed to tweak how has_value is set for null table column count mismatches.